@@ -4,13 +4,13 @@ build-backend = "flit_core.buildapi"
 
 [project]
 name = "tomli"
-version = "1.2.2"  # DO NOT EDIT THIS LINE MANUALLY. LET bump2version UTILITY DO IT
+version = "2.0.1"  # DO NOT EDIT THIS LINE MANUALLY. LET bump2version UTILITY DO IT
 description = "A lil' TOML parser"
 authors = [
     { name = "Taneli Hukkinen", email = "hukkin@users.noreply.github.com" },
 ]
 license = { file = "LICENSE" }
-requires-python = ">=3.6"
+requires-python = ">=3.7"
 readme = "README.md"
 classifiers = [
     "License :: OSI Approved :: MIT License",
@@ -18,7 +18,6 @@ classifiers = [
     "Operating System :: Microsoft :: Windows",
     "Operating System :: POSIX :: Linux",
     "Programming Language :: Python :: 3 :: Only",
-    "Programming Language :: Python :: 3.6",
     "Programming Language :: Python :: 3.7",
     "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
@@ -35,16 +34,6 @@ keywords = ["toml"]
 "Changelog" = "https://github.com/hukkin/tomli/blob/master/CHANGELOG.md"
 
 
-[tool.flit.sdist]
-exclude = [
-    "tests/",
-    "benchmark/",
-    "fuzzer/",
-    ".*",
-    "CHANGELOG.md",
-]
-
-
 [tool.isort]
 # Force imports to be sorted by module, independent of import type
 force_sort_within_sections = true
@@ -58,42 +47,26 @@ known_first_party = ["tomli", "tests"]
 profile = "black"
 
 
-[tool.pytest.ini_options]
-addopts = "--strict-markers --strict-config"
-xfail_strict = true
-
-
 [tool.tox]
 legacy_tox_ini = '''
 [tox]
-# Only run pytest envs when no args given to tox
-envlist = py{36,37,38,39,310}
+# Only run unittest envs when no args given to tox
+envlist = py{37,38,39,310}
 isolated_build = True
 
-[testenv:py{36,37,38,39,310}]
-description = run tests against unpackaged source
-skip_install = True
-deps = -r tests/requirements.txt
+[testenv:py{37,38,39,310}]
+description = run tests against a built package
 commands =
-    # Use 'python -m pytest' to add CWD to sys.path
-    python -m pytest {posargs}
-
-[testenv:py{36,37,38,39,310}-package]
-description = run tests against a built package (can fail, in theory, if test dependencies need a tomli version incompatible with local state)
-deps = -r tests/requirements.txt
-commands =
-    pytest --import-mode=append {posargs}
+    python -m unittest {posargs}
 
 [testenv:profile]
 description = run profiler (use e.g. `firefox .tox/prof/combined.svg` to open)
-skip_install = True
 setenv =
     PROFILER_ITERATIONS=1000
 deps =
-    -r tests/requirements.txt
-    pytest-profiling
+    -r profiler/requirements.txt
 commands =
-    pytest tests/test_for_profiler.py --profile-svg --pstats-dir "{toxworkdir}/prof"
+    pytest profiler/test_for_profiler.py --profile-svg --pstats-dir "{toxworkdir}/prof"
     python -c 'import pathlib; print("profiler svg output under file://\{0\}".format(pathlib.Path(r"{toxworkdir}") / "prof" / "combined.svg"))'
 
 [testenv:pre-commit]
@@ -140,15 +113,15 @@ commands =
 
 [tool.coverage.run]
 branch = true
-omit = ["tests/*", "*/site-packages/*"]
+source = ['tomli']
 
 [tool.coverage.report]
 # Regexes for lines to exclude from consideration
 exclude_lines = [
-    # Have to re-enable the standard pragma
-    "pragma: no cover",
+    # Re-enable the standard pragma (with extra strictness)
+    '# pragma: no cover\b',
     # Code for static type checkers
-    "if TYPE_CHECKING:",
+    'if TYPE_CHECKING:',
     # Scripts
     'if __name__ == .__main__.:',
 ]
